Move dependencies around, trying to speed import of ReconsPlan
