+++ phaser/types.py
@@ -5,7 +5,7 @@ import typing as t
 import numpy
 import pane
 from pane.converters import Converter, make_converter, ConverterHandlers, ErrorNode
-from pane.annotations import ConvertAnnotation
+from pane.annotations import ConvertAnnotation, Condition
 from pane.errors import ParseInterrupt, WrongTypeError
 from pane.util import pluralize, list_phrase
 from typing_extensions import Self
@@ -120,6 +120,90 @@ class SliceTotal(Dataclass):
         return [self.total_thickness / self.n] * self.n
 
 Slices: t.TypeAlias = t.Union[SliceList, SliceStep, SliceTotal]
+
+
+class ComplexCartesian(pane.PaneBase, kw_only=True):
+    re: float
+    im: float = 0.0
+
+    def __complex__(self) -> complex:
+        return complex(self.re, self.im)
+
+class ComplexPolar(pane.PaneBase, kw_only=True):
+    mag: float
+    angle: float = 0.0  # degrees
+
+    def __complex__(self) -> complex:
+        theta = numpy.deg2rad(self.angle)
+        return self.mag * complex(numpy.cos(theta), numpy.sin(theta))
+
+
+class Krivanek(pane.PaneBase):
+    n: int
+    m: int
+    scale_factor: float = 1.0
+
+    def __post_init__(self):
+        if (
+            self.n < 0 or self.m < 0 or
+            self.m > self.n + 1 or
+            self.m % 2 + self.n % 2 != 1
+        ):
+            raise ValueError(f"Invalid Krivanek aberration n={self.n} m={self.m}")
+
+    @staticmethod
+    def from_known(s: str) -> 'Krivanek':
+        try:
+            return _KNOWN_ABERRATIONS[s.lower()]
+        except (KeyError, TypeError):
+            raise ValueError(f"Unknown aberration '{s}'") from None
+
+class KrivanekComplex(Krivanek, kw_only=True):
+    val: complex
+
+    def __complex__(self) -> complex:
+        return self.val
+
+class KrivanekCartesian(Krivanek, ComplexCartesian, kw_only=True):
+    ...
+
+class KrivanekPolar(Krivanek, ComplexPolar, kw_only=True):
+    ...
+
+
+_KNOWN_ABERRATIONS: t.Dict[str, Krivanek] = {
+    'c1': Krivanek.make_unchecked(1, 0),
+    'a1': Krivanek.make_unchecked(1, 2),
+    'b2': Krivanek.make_unchecked(2, 1, 3.0),  # C_21 = 3*B2
+    'a2': Krivanek.make_unchecked(2, 3),
+    'c3': Krivanek.make_unchecked(3, 0),
+    's3': Krivanek.make_unchecked(3, 2, 3.0),  # C_32 = 3*S3
+    'a3': Krivanek.make_unchecked(3, 4),
+    'b4': Krivanek.make_unchecked(4, 1, 4.0),  # C_41 = 4*B4
+    'd4': Krivanek.make_unchecked(4, 3, 4.0),  # C_43 = 4*D4
+    'a4': Krivanek.make_unchecked(4, 5),
+    'c5': Krivanek.make_unchecked(5, 0),
+}
+
+KnownAberration: t.TypeAlias = t.Annotated[str, Condition(
+    lambda s: s.lower() in _KNOWN_ABERRATIONS,
+    'known aberration',
+    lambda exp, plural: pluralize('known aberration', plural)
+)]
+
+Aberration: t.TypeAlias = t.Union[
+    t.Dict[KnownAberration, t.Union[complex, ComplexCartesian, ComplexPolar]],
+    KrivanekComplex, KrivanekCartesian, KrivanekPolar,
+]
+
+def process_aberrations(aberrations: t.Iterable[Aberration]) -> t.Iterator[KrivanekComplex]:
+    for ab in aberrations:
+        if isinstance(ab, dict):
+            for known, val in ab.items():
+                ty = Krivanek.from_known(known)
+                yield KrivanekComplex(ty.n, ty.m, val=ty.scale_factor * complex(val))
+        else:
+            yield KrivanekComplex(ab.n, ab.m, val=complex(ab))
 
 
 class SimpleFlag(Dataclass):
@@ -302,24 +386,10 @@ class _ReconsVarsConverter(Converter[t.FrozenSet[ReconsVar]]):
         return self.inner.collect_errors(val)
 
 
-class ComplexCartesian(pane.PaneBase, kw_only=True):
-    re: float
-    im: float = 0.0
-
-    def __complex__(self) -> complex:
-        return complex(self.re, self.im)
-
-
-class ComplexPolar(pane.PaneBase, kw_only=True):
-    mag: float
-    angle: float = 0.0  # degrees
-
-    def __complex__(self) -> complex:
-        theta = numpy.deg2rad(self.angle)
-        return self.mag * complex(numpy.cos(theta), numpy.sin(theta))
-
-
 __all__ = [
     'BackendName', 'Dataclass', 'Slices', 'Flag',
-    'process_flag', 'flag_any_true',
+    'ComplexCartesian', 'ComplexPolar',
+    'Krivanek', 'KrivanekComplex', 'KrivanekCartesian',
+    'KrivanekPolar', 'KnownAberration', 'Aberration',
+    'process_aberrations', 'process_flag', 'flag_any_true',
 ]
